+++ src/components/shop/shop-menu-screen.js
@@ -6,11 +6,10 @@
  */
 import { Screen } from "../../core/screen.js";
 import RexUIPlugin from "../../../lib/rexuiplugin.min.js";
-import { getMetrics, onScaleChange } from "../../core/scaler.js";
+import { onScaleChange } from "../../core/scaler.js";
 import { createTitles } from "../../core/titles.js";
 import { createBalance } from "./balance-ui.js";
 import { createMenu } from "./menu.js";
-import { getSafeArea, getXPos, getYPos, getScaleFactor } from "./shop-layout.js";
 
 export class ShopMenu extends Screen {
     preload() {
@@ -36,13 +35,7 @@ export class ShopMenu extends Screen {
     }
 
     resize() {
-        const metrics = getMetrics();
         this.menu.resize();
-        const safeArea = getSafeArea(this.layout);
-        this.balance.setScale(getScaleFactor({ metrics, container: this.balance, safeArea }));
-        this.balance.setPosition(
-            getXPos(this.balance, safeArea, this.config.listPadding.x),
-            getYPos(metrics, safeArea),
-        );
+        this.balance.resize();
     }
 }
